JavaFX and Maven Injected
