--- src/main/Main.java
@@ -1,14 +1,0 @@
-package main;
-
-import main.characters.Characters;
-import main.characters.Elarix;
-import main.characters.Soldier;
-import main.game.GameLogic;
-import main.game.Story;
-
-public class Main {
-
-    public static void main(String[] args) {
-        GameLogic.startGame();
-    }
-}
