--- src/main/characters/Aliens.java
@@ -1,16 +1,0 @@
-package main.characters;
-
-public class Aliens extends Characters {
-
-
-    public Aliens(String name, int health, int armor, int damage, int xp) {
-        super(name, health, armor, damage, xp);
-    }
-
-
-    //unique methods
-
-    void heal(){}
-    void mutate(){}
-
-}
