--- src/main/characters/Elarix.java
@@ -1,9 +1,0 @@
-package main.characters;
-
-public class Elarix extends Aliens {
-    public Elarix() {
-        super("Elarix", 30, 1, 2, 1);
-    }
-
-    //skills
-}
