--- src/main/characters/Humans.java
@@ -1,13 +1,0 @@
-package main.characters;
-
-public class Humans extends Characters{
-
-
-
-    public Humans(String name,int health, int armor,int damage,int xp) {
-        super(name,health,armor,damage,xp);
-    }
-
-}
-
-
