--- src/main/game/GameExceptions.java
@@ -1,8 +1,0 @@
-package main.game;
-
-public class GameExceptions extends Exception{
-
-    public GameExceptions(String message){
-        super(message);
-    }
-}
