+++ src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module org.zimgo.hellofx {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+
+    opens org.zimgo.aliensvshumans to javafx.fxml;
+    exports org.zimgo.aliensvshumans;
+}
