+++ src/main/java/org/zimgo/aliensvshumans/MainController.java
@@ -1,0 +1,14 @@
+package org.zimgo.aliensvshumans;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class MainController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
