+++ src/main/java/org/zimgo/aliensvshumans/characters/Aliens.java
@@ -1,0 +1,16 @@
+package org.zimgo.aliensvshumans.characters;
+
+public class Aliens extends Characters {
+
+
+    public Aliens(String name, int health, int armor, int damage, int xp) {
+        super(name, health, armor, damage, xp);
+    }
+
+
+    //unique methods
+
+    void heal(){}
+    void mutate(){}
+
+}
