+++ src/main/java/org/zimgo/aliensvshumans/characters/Elarix.java
@@ -1,0 +1,9 @@
+package org.zimgo.aliensvshumans.characters;
+
+public class Elarix extends Aliens {
+    public Elarix() {
+        super("Elarix", 30, 1, 2, 1);
+    }
+
+    //skills
+}
