+++ src/main/java/org/zimgo/aliensvshumans/characters/Humans.java
@@ -1,0 +1,13 @@
+package org.zimgo.aliensvshumans.characters;
+
+public class Humans extends Characters{
+
+
+
+    public Humans(String name,int health, int armor,int damage,int xp) {
+        super(name,health,armor,damage,xp);
+    }
+
+}
+
+
