+++ src/main/java/org/zimgo/aliensvshumans/game/GameExceptions.java
@@ -1,0 +1,8 @@
+package org.zimgo.aliensvshumans.game;
+
+public class GameExceptions extends Exception{
+
+    public GameExceptions(String message){
+        super(message);
+    }
+}
