+++ src/test/java/test/testGame/TestGameLogic.java
@@ -1,0 +1,28 @@
+//package test.testGame;
+//
+//
+//import org.zimgo.aliensvshumans.characters.Characters;
+//import org.zimgo.aliensvshumans.characters.Elarix;
+//import org.zimgo.aliensvshumans.characters.Soldier;
+//import org.zimgo.aliensvshumans.game.GameLogic;
+//
+//
+//
+//
+//class TestGameLogic{
+//    Characters char1 = new Soldier("Ry");
+//    Characters char2 = new Elarix();
+//
+//    @Test
+//    void testShowCharDetails(){
+//        GameLogic.showCharacterDetails(char1);
+//        GameLogic.showCharacterDetails(char2);
+//    }
+//
+//    @Test
+//    void testBattleSystem(){
+//        GameLogic.battleSystem(char1,char2);
+//    }
+//
+//
+//}
